areSame is now part of the API
