+++ src/main/java/mods/eln/Eln.java
@@ -10,6 +10,7 @@ import cpw.mods.fml.common.network.simpleimpl.SimpleNetworkWrapper;
 import cpw.mods.fml.common.registry.EntityRegistry;
 import cpw.mods.fml.common.registry.GameRegistry;
 import cpw.mods.fml.relauncher.Side;
+import mods.eln.api.Utilities;
 import mods.eln.api.recipe.Recipe;
 import mods.eln.api.recipe.RecipesList;
 import mods.eln.packets.*;
@@ -944,7 +945,7 @@ public class Eln {
                 IRecipe r = (IRecipe) o;
                 if (r.getRecipeOutput() == null)
                     continue;
-                if (Utils.areSame(stack, r.getRecipeOutput()))
+                if (Utilities.areSame(stack, r.getRecipeOutput()))
                     return true;
             }
         }
