+++ src/main/java/mods/eln/api/Utilities.java
@@ -1,6 +1,7 @@
 package mods.eln.api;
 
 import net.minecraft.item.ItemStack;
+import net.minecraftforge.oredict.OreDictionary;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
@@ -10,22 +11,6 @@ import java.lang.reflect.Modifier;
  * Created by Gregory Maddra on 2016-11-18.
  */
 public class Utilities {
-
-    public static boolean areSame(ItemStack input, ItemStack output){
-        try {
-            Class<?> Utils = Class.forName("mods.eln.misc.Utils");
-            return (Boolean) Utils.getDeclaredMethod("areSame", ItemStack.class, ItemStack.class).invoke(null, input, output);
-        } catch (ClassNotFoundException e) {
-            e.printStackTrace();
-        } catch (NoSuchMethodException e) {
-            e.printStackTrace();
-        } catch (InvocationTargetException e) {
-            e.printStackTrace();
-        } catch (IllegalAccessException e) {
-            e.printStackTrace();
-        }
-        return false;
-    }
 
     public static void makeModifiable(Field field){
         field.setAccessible(true);
@@ -41,5 +26,21 @@ public class Utilities {
             e.printStackTrace();
         }
 
+    }
+
+    public static boolean areSame(ItemStack stack, ItemStack output) {
+        try {
+            if (stack.getItem() == output.getItem() && stack.getItemDamage() == output.getItemDamage()) return true;
+            int[] stackIds = OreDictionary.getOreIDs(stack);
+            int[] outputIds = OreDictionary.getOreIDs(output);
+            // System.out.println(Arrays.toString(stackIds) + "   " + Arrays.toString(outputIds));
+            for (int i : outputIds) {
+                for (int j : stackIds) {
+                    if (i == j) return true;
+                }
+            }
+        } catch (Exception e) {
+        }
+        return false;
     }
 }
