+++ src/main/java/mods/eln/api/recipe/Recipe.java
@@ -4,7 +4,7 @@ import net.minecraft.item.ItemStack;
 
 import java.util.ArrayList;
 
-import mods.eln.misc.Utils;
+import mods.eln.api.Utilities;
 import net.minecraft.item.ItemStack;
 
 public class Recipe {
@@ -25,10 +25,18 @@ public class Recipe {
         this.energy = energy;
     }
 
-    public boolean canBeCraftedBy(ItemStack stack) {
-        if (stack == null) return false;
-        return input.stackSize <= stack.stackSize && Utils.areSame(stack, input);
-    }
+	public boolean canBeCraftedBy(ItemStack stack) {
+		if(stack == null) return false;
+		return input.stackSize <= stack.stackSize && Utilities.areSame(stack, input);
+	}
+	
+	public ItemStack[] getOutputCopy() {
+		ItemStack[] cpy = new ItemStack[output.length];
+		for(int idx = 0; idx<output.length; idx++) {
+			cpy[idx] = output[idx].copy();
+		}
+		return cpy;
+	}
 
     public ItemStack[] getOutputCopy() {
         ItemStack[] cpy = new ItemStack[output.length];
